+++ lib/analytics/analytics_proxy.dart
@@ -3,74 +3,75 @@ import 'base_analytics.dart';
 /// A common default analytic proxy to use to capture analytics events
 class AnalyticsProxy extends BaseAnalytics {
   final List<BaseAnalytics> _analyticList;
+  bool enable;
 
   /// Pass a list of analytics implementations of [BaseAnalytics]
   /// If list is empty no analytics will be sent out to any element
-  AnalyticsProxy(this._analyticList, {required bool enable}) : super() {
-    if (enable) initialize();
-  }
+  AnalyticsProxy(this._analyticList, {required this.enable}) : super();
 
   @override
-  logEvent(String? name, List<Type>? sendTo, {Map<String, dynamic>? properties = const {}}) {
+  Future<void> logEvent(String? name, List<Type>? sendTo,
+      {Map<String, dynamic>? properties = const {}}) async {
     if (name != null && name.isNotEmpty) {
       if (isInitialized == true) {
-        _analyticList.forEach((BaseAnalytics analyticPlatform) {
-          analyticPlatform.logEvent(name, sendTo, properties: properties);
+        await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+          await analyticPlatform.logEvent(name, sendTo, properties: properties);
         });
       }
     }
   }
 
-  /// this will automatically be called on the constructor
+  ///  will need to be called before any analytics are sent out
   @override
-  void initialize() {
-    if (!isInitialized) {
+  Future<void> initialize() async {
+    if (!isInitialized && enable) {
       isInitialized = true;
-      _analyticList.forEach((BaseAnalytics analyticPlatform) {
-        analyticPlatform.initialize();
+      await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+        await analyticPlatform.initialize();
       });
     }
   }
 
   @override
-  logPageName(String? pageName, List<Type> sendTo, {Map<String, dynamic> properties = const {}}) {
+  Future<void> logPageName(String? pageName, List<Type> sendTo,
+      {Map<String, dynamic> properties = const {}}) async {
     if (pageName == null || pageName.isEmpty) return;
-    _analyticList.forEach((BaseAnalytics analyticPlatform) {
-      analyticPlatform.logPageName(pageName, sendTo, properties: properties);
+    await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+      await analyticPlatform.logPageName(pageName, sendTo, properties: properties);
     });
   }
 
   @override
-  void identifyUser(String? uid) {
-    _analyticList.forEach((BaseAnalytics analyticPlatform) {
-      analyticPlatform.identifyUser(uid);
+  Future<void> identifyUser(String? uid) async {
+    await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+      await analyticPlatform.identifyUser(uid);
     });
   }
 
   @override
-  void setUserProperties(Map<String, dynamic> properties) {
-    _analyticList.forEach((BaseAnalytics analyticPlatform) {
-      analyticPlatform.setUserProperties(properties);
+  Future<void> setUserProperties(Map<String, dynamic> properties) async {
+    await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+      await analyticPlatform.setUserProperties(properties);
     });
   }
 
   @override
-  void setUserPropertiesWithOptions(Map<String, dynamic> properties, String analyticOption) {
-    _analyticList.forEach((BaseAnalytics analyticPlatform) {
-      analyticPlatform.setUserPropertiesWithOptions(properties, analyticOption);
+  Future<void> setUserPropertiesWithOptions(Map<String, dynamic> properties, String analyticOption) async {
+    await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+      await analyticPlatform.setUserPropertiesWithOptions(properties, analyticOption);
     });
   }
 
-  dispose() {
-    _analyticList.forEach((BaseAnalytics analyticPlatform) {
-      analyticPlatform.dispose();
+  Future<void> dispose() async {
+    await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+      await analyticPlatform.dispose();
     });
   }
 
   @override
-  isSensitiveScreen(bool isSensitive, List<Type> sendTo, {Map<String, dynamic> properties = const {}}) {
-    _analyticList.forEach((BaseAnalytics analyticPlatform) {
-      analyticPlatform.isSensitiveScreen(isSensitive, sendTo, properties: properties);
+  isSensitiveScreen(bool isSensitive, List<Type> sendTo, {Map<String, dynamic> properties = const {}}) async {
+    await Future.forEach(_analyticList, (BaseAnalytics analyticPlatform) async {
+      await analyticPlatform.isSensitiveScreen(isSensitive, sendTo, properties: properties);
     });
   }
 }