+++ lib/analytics/base_analytics.dart
@@ -4,26 +4,26 @@ abstract class BaseAnalytics {
 
   BaseAnalytics();
 
-  logEvent(String name, List<Type>? sendTo, {Map<String, dynamic>? properties = const {}});
+  Future<void> logEvent(String name, List<Type>? sendTo, {Map<String, dynamic>? properties = const {}});
 
-  initialize();
+  Future<void> initialize();
 
-  logPageName(String pageName, List<Type> sendTo, {Map<String, dynamic> properties = const {}});
+  Future<void> logPageName(String pageName, List<Type> sendTo, {Map<String, dynamic> properties = const {}});
 
   isSensitiveScreen(bool isSensitive, List<Type> sendTo, {Map<String, dynamic> properties = const {}});
 
   bool isInitialized = false;
 
-  void identifyUser(String? uid);
+  Future<void> identifyUser(String? uid);
 
   bool shouldSendToThisPlatform(List<Type>? allowedTypes) {
     if (allowedTypes == null || allowedTypes.isEmpty) return false;
     return allowedTypes.contains(this.runtimeType);
   }
 
-  void setUserProperties(Map<String, dynamic> properties);
+  Future<void> setUserProperties(Map<String, dynamic> properties);
 
-  void setUserPropertiesWithOptions(Map<String, dynamic> properties, String analyticOption);
+  Future<void> setUserPropertiesWithOptions(Map<String, dynamic> properties, String analyticOption);
 
-  dispose();
+  Future<void> dispose();
 }